test: chunking fixtures live once in the root conftest
The MockMemoryManager and the forced_free_mem/low_memory/
low_mem_solver/unchunking_solver/chunked_solved_solver/
unchunked_solved_solver fixtures were defined identically in
batchsolving/arrays/conftest.py and integrated_numerical_tests/
conftest.py; both subdirectory conftests are removed and the single
copy (with the arrays copy's chunk-sizing arithmetic comment) lives
in tests/conftest.py.
